Add RSS settings to statPlugin control, refactor
This commit is a fairly significant refactoring of the control side of
the statPlugin. The primary addition is making the RSS credentials for
the STAT+ feed configurable, but in support of that there’s a major
refactor and abstraction so that individual forms can be added to the
page easily.

There were also a few changes necessary over on the widget side to
fetch/cache the feed credentials, as well as some minor cleanup and
marking async functions as async since they eventually return a Promise.